move cli functions to separate import
